backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
pillow
pytesseract
pdfplumber
nltk
spacy
openai
faiss-cpu
python-multipart
python-dotenv
requests
